Crear fragmento de Registrar Vehiculo Modificar framento de registrar vehiculo y registrar reparacion como dialogs
