+++ .idea/misc.xml
@@ -7,17 +7,17 @@
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/font/amaranth_bold.xml" value="0.3927083333333333" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/activity_main.xml" value="0.264" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/activity_splash_screen.xml" value="0.3625" />
-        <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/dialog_registar_taller.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/fragmento_hist_reparaciones.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/fragmento_talleres.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/fragmento_vehiculos.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/list_item_hist_reparaciones.xml" value="0.4" />
-        <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/list_item_talleres.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/layout/list_item_vehiculos.xml" value="0.3625" />
         <entry key="..\:/Users/GHTDEVELOPER/Documents/proyect_android/CarCare/app/src/main/res/menu/menu_bottom_navegation.xml" value="0.264" />
         <entry key="app/src/main/res/layout/activity_main.xml" value="0.3203125" />
         <entry key="app/src/main/res/layout/activity_splash_screen.xml" value="0.32760416666666664" />
-        <entry key="app/src/main/res/layout/fragment_registrar_reparacion.xml" value="0.1650943396226415" />
+        <entry key="app/src/main/res/layout/dialog_registar_taller.xml" value="0.16530797101449277" />
+        <entry key="app/src/main/res/layout/fragment_registrar_reparacion.xml" value="0.1" />
+        <entry key="app/src/main/res/layout/fragment_registrar_vehiculo.xml" value="0.2109375" />
         <entry key="app/src/main/res/layout/fragmento_hist_reparaciones.xml" value="0.32760416666666664" />
         <entry key="app/src/main/res/layout/fragmento_talleres.xml" value="0.33" />
         <entry key="app/src/main/res/layout/fragmento_vehiculos.xml" value="0.32760416666666664" />
