+++ app/src/main/java/es/usj/mastertsa/carcare/ui/hist_reparaciones/FragmentoHistReparaciones.kt
@@ -1,5 +1,6 @@
 package es.usj.mastertsa.carcare.ui.hist_reparaciones
 
+import android.app.AlertDialog
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
@@ -21,9 +22,7 @@ private const val ARG_PARAM2 = "param2"
  * create an instance of this fragment.
  */
 class FragmentoHistReparaciones : Fragment() {
-    // TODO: Rename and change types of parameters
-    private var param1: String? = null
-    private var param2: String? = null
+
 
     private lateinit var bindings : FragmentoHistReparacionesBinding
 
@@ -31,10 +30,7 @@ class FragmentoHistReparaciones : Fragment() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        arguments?.let {
-            param1 = it.getString(ARG_PARAM1)
-            param2 = it.getString(ARG_PARAM2)
-        }
+
     }
 
     override fun onCreateView(
@@ -59,9 +55,18 @@ class FragmentoHistReparaciones : Fragment() {
 
     private fun loadRegistrarReparacion()
     {
-        registrarReparacion = RegistrarReparacion()
-        activity?.supportFragmentManager?.beginTransaction()
-            ?.add(R.id.fragment_container_view,registrarReparacion)?.commit()
+//        registrarReparacion = RegistrarReparacion()
+//        activity?.supportFragmentManager?.beginTransaction()
+//            ?.add(R.id.fragment_container_view,registrarReparacion)?.commit()
+
+        val dialogBuilder = AlertDialog.Builder(requireContext())
+        val inflater = this.layoutInflater
+        val dialogView = inflater.inflate(R.layout.fragment_registrar_reparacion,null)
+        dialogBuilder.setView(dialogView)
+
+        val alertDialog = dialogBuilder.create()
+        alertDialog.setTitle("Registrar Reparación")
+        alertDialog.show()
     }//Fin de la funcion loadRegistrarReparacion
