+++ app/src/main/res/values/strings.xml
@@ -26,6 +26,8 @@
     <string name="text_nombre_taller">Nombre</string>
     <string name="text_registrar">Registrar</string>
     <string name="text_cancelar">Cancelar</string>
+    <string name="dummy_button">Dummy Button</string>
+    <string name="dummy_content">DUMMY\nCONTENT</string>
 
 
 </resources>
